requirements.txt
# Invenio requirements.
MySQL-python==1.2.5
rdflib==2.4.2
reportlab==2.5
python-dateutil<=1.9999
python-magic==0.4.2
http://www.reportlab.com/ftp/pyRXP-1.16-daily-unix.tar.gz
numpy==1.8.1
lxml==3.1.2
mechanize==0.2.5
PyStemmer==1.3.0
Python-Levenshtein==0.11.2
feedparser==5.1.3
BeautifulSoup==3.2.1
beautifulsoup4==4.1.3
python-twitter==0.8.7
msgpack-python==0.3.0
pyparsing==1.5.6
requests==2.2.0
PyPDF2==1.19
rauth==0.6.2
unidecode==0.04.14
python-openid==2.2.5
qrcode==4.0.4
pillow==2.3.0
jinja2==2.7.2
redis==2.9.0
nydus==0.10.6
Cerberus==0.5
HTMLParser==0.0.2
h5py==2.3.1
unittest2==0.5.1
invenio-query-parser==0.1